requirements.txt
streamlit>=1.30
python-pptx>=0.6.23
Pillow>=10.0
